+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
   - sqlobject's `EmptyHandling` enum backported to core for invocations of `SqlStatement.bindList`
   - installPlugin skips duplicate JdbiPlugins (according to Object.equals)
   - KotlinSqlObjectPlugin will install forgotten SqlObjectPlugin for you
+  - KotlinMapper respects `@PropagateNull`
 - New Beta Features
   - added `register` methods for qualified factories on `Configurable`,
     `ColumnMappers`, and `ArgumentFactories`